move `DEFAULT_CATEGORICAL_SLOT_VALUE` constant to `shared`
