+++ rasa/core/constants.py
@@ -69,7 +69,6 @@ UTTER_PREFIX = "utter_"
 RESPOND_PREFIX = "respond_"
 
 DIALOGUE = "dialogue"
-DEFAULT_CATEGORICAL_SLOT_VALUE = "__other__"
 
 # RabbitMQ message property header added to events published using `rasa export`
 RASA_EXPORT_PROCESS_ID_HEADER_NAME = "rasa-export-process-id"
