+++ rasa/shared/core/constants.py
@@ -1,0 +1,1 @@
+DEFAULT_CATEGORICAL_SLOT_VALUE = "__other__"
