+++ rasa/shared/core/slots.py
@@ -3,7 +3,7 @@ from typing import Any, Dict, List, NoReturn, Optional, Text, Type
 
 from rasa.core import utils
 from typing import Any, Dict, List, NoReturn, Optional, Text, Type
-from rasa.core.constants import DEFAULT_CATEGORICAL_SLOT_VALUE
+from rasa.shared.core.constants import DEFAULT_CATEGORICAL_SLOT_VALUE
 from rasa.utils.common import class_from_module_path, raise_warning
 
 logger = logging.getLogger(__name__)
